+++ CHANGELOG.md
@@ -32,11 +32,13 @@ modified Semantic Versioning scheme. See the "Versioning scheme" section of the
 - RIG-56: Added a configuration form to toggle content types for the eCMS API. 
 - RIG-57: Added the ecms_api_publisher module and custom entity.
 - RIG-60: Added the ecms_api_publisher module installation steps.
+- RIG-22: Added the ecms_workflow module to automatically assign content types to default workflow and set permissions.
 
 ### Changed
 - RIG-23: Changed from OIDC generic to Windows AAD for authentication.
 - Disabled xdebug by default in the develop.sh script.
 - RIG-37: Made ECMS custom theme the default.
+- RIG-22: Enable Moderation Dashboard module by default.
 
 ### Deprecated